refactor: better separate between "framework" and "app" code
