+++ src/app/auth.ts
@@ -2,14 +2,14 @@ import { APIError, FilenSDKConfig } from "@filen/sdk"
 import fs from "node:fs"
 import { exists } from "./util/util"
 import path from "path"
-import { wrapRedTerminalText } from "./interface/util"
+import { wrapRedTerminalText } from "../framework/util"
 import { ANONYMOUS_SDK_CONFIG } from "./constants"
 import crypto from "node:crypto"
-import { isRunningAsContainer } from "./buildInfo"
+import { isRunningAsContainer } from "../buildInfo"
 import dedent from "dedent"
 import { f, X } from "./app"
-import { FeatureContext, FeatureGroup } from "./framework/features"
-import { App } from "./framework/app"
+import { FeatureContext, FeatureGroup } from "../framework/features"
+import { App } from "../framework/app"
 
 const authConfigFileName = ".filen-cli-auth-config"
 const keepMeLoggedInFile = (app: App<X>) => path.join(app.dataDir, ".filen-cli-keep-me-logged-in")
