+++ src/app/featureInterfaces/publicLinksInterface.ts
@@ -1,8 +1,8 @@
 import FilenSDK, { PublicLinkExpiration } from "@filen/sdk"
-import { formatTable, formatTimestamp } from "../interface/util"
-import { getItemPaths } from "../util/util"
-import { FeatureGroup } from "../framework/features"
+import { formatTable, formatTimestamp } from "../../framework/util"
+import { FeatureGroup } from "../../framework/features"
 import { f, X } from "../app"
+import { getItemPaths } from "../util/util"
 
 async function getPublicLinkStatus(filen: FilenSDK, type: "file" | "directory", uuid: string, fileKey: string | undefined): Promise<{
 	type: "file" | "directory"
