+++ src/app/featureInterfaces/s3Interface.ts
