+++ src/index.ts
@@ -3,9 +3,9 @@
 import { checkInjectedBuildInfo } from "./buildInfo"
 import { read } from "read"
 import { CompleterResult } from "readline"
-import { wrapRedTerminalText } from "./interface/util"
+import { wrapRedTerminalText } from "./framework/util"
 import { InterfaceAdapter } from "./framework/app"
-import { app } from "./app"
+import { app } from "./app/app"
 
 if (!checkInjectedBuildInfo()) {
 	console.error("Build info not injected correctly!")
